refactor(dialogs): extract ReleaseNotesContent shared widget
Pull the single-release rendering ("_ReleaseCard" + "_markdownConfig")
out of new_version_dialog.dart into a standalone widget that the
upcoming CurrentReleaseDialog can reuse. NewVersionDialog now composes
it; visual and behavioural output is unchanged.
