+++ lib/widgets/dialogs/new_version_dialog.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:markdown_widget/markdown_widget.dart';
+
 import 'package:genshin_impact_wish_gacha_analyzer/l10n/generated/app_localizations.dart';
 import 'package:genshin_impact_wish_gacha_analyzer/state/app_release.dart';
 import 'package:genshin_impact_wish_gacha_analyzer/theme/tokens.dart';
 import 'package:genshin_impact_wish_gacha_analyzer/widgets/app_link.dart';
 import 'package:genshin_impact_wish_gacha_analyzer/widgets/dialogs/app_dialog.dart';
-import 'package:genshin_impact_wish_gacha_analyzer/widgets/relative_time_text.dart';
+import 'package:genshin_impact_wish_gacha_analyzer/widgets/dialogs/release_notes_content.dart';
 
 /// 新版本通知 dialog，列出自上次版本以來的所有 release notes。
 class NewVersionDialog extends ConsumerWidget {
@@ -37,7 +37,7 @@ class NewVersionDialog extends ConsumerWidget {
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: [
           for (int i = 0; i < releases.length; i++) ...[
-            _ReleaseCard(release: releases[i], l: l),
+            ReleaseNotesContent(release: releases[i]),
             if (i < releases.length - 1) const SizedBox(height: AppSpacing.m),
           ],
         ],
@@ -65,72 +65,6 @@ class NewVersionDialog extends ConsumerWidget {
           },
         ),
       ],
-    );
-  }
-}
-
-/// 依 [theme] 亮暗模式建立 markdown 渲染設定，統一連結樣式。
-MarkdownConfig _markdownConfig(ThemeData theme) {
-  final base = theme.brightness == Brightness.dark
-      ? MarkdownConfig.darkConfig
-      : MarkdownConfig.defaultConfig;
-  return base.copy(
-    configs: [LinkConfig(style: TextStyle(color: linkBaseColor(theme)))],
-  );
-}
-
-/// 單一 release 的卡片，顯示版本號、發布時間與 markdown release notes。
-class _ReleaseCard extends StatelessWidget {
-  const _ReleaseCard({required this.release, required this.l});
-
-  /// 該卡片對應的 release 資料。
-  final AppRelease release;
-
-  /// 在地化資源，用於時間格式文案。
-  final AppLocalizations l;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final tokens = theme.gacha;
-
-    return Card(
-      margin: EdgeInsets.zero,
-      child: Padding(
-        padding: const EdgeInsets.all(AppSpacing.m),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              crossAxisAlignment: CrossAxisAlignment.baseline,
-              textBaseline: TextBaseline.alphabetic,
-              children: [
-                Text(
-                  release.tagName,
-                  style: theme.textTheme.titleLarge?.copyWith(
-                    color: theme.colorScheme.primary,
-                    fontWeight: FontWeight.w700,
-                  ),
-                ),
-                const SizedBox(width: AppSpacing.s),
-                RelativeTimeText(
-                  time: release.publishedAt,
-                  templateBuilder: l.updateReleasedAt,
-                  useDateOnly: true,
-                  style: theme.textTheme.bodySmall?.copyWith(
-                    color: tokens.textMuted,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: AppSpacing.s),
-            Divider(height: 1, color: tokens.borderSubtle),
-            const SizedBox(height: AppSpacing.s),
-            if (release.body.isNotEmpty)
-              MarkdownBlock(data: release.body, config: _markdownConfig(theme)),
-          ],
-        ),
-      ),
     );
   }
 }
